refactor: 파사드 추가 및 concert service, repository 책임 분리
