+++ src/main/java/kr/hhplus/be/server/concert/domain/repository/ConcertScheduleRepository.java
@@ -1,5 +1,12 @@
 package kr.hhplus.be.server.concert.domain.repository;
 
+import kr.hhplus.be.server.concert.domain.entity.ConcertSchedule;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
 public interface ConcertScheduleRepository {
-//    List<ConcertSchedule> findAvailableConcertSchedules(long concertId, LocalDateTime currentDate);
+
+    List<ConcertSchedule> findAvailableConcertSchedules(long concertId, LocalDateTime currentDate);
+
 }
