+++ src/main/java/kr/hhplus/be/server/concert/infra/ConcertScheduleRepositoryImpl.java
@@ -1,0 +1,24 @@
+package kr.hhplus.be.server.concert.infra;
+
+import kr.hhplus.be.server.concert.domain.entity.ConcertSchedule;
+import kr.hhplus.be.server.concert.domain.repository.ConcertScheduleRepository;
+import kr.hhplus.be.server.concert.infra.orm.ConcertScheduleJPARepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+
+@Repository
+@RequiredArgsConstructor
+public class ConcertScheduleRepositoryImpl implements ConcertScheduleRepository {
+
+    private final ConcertScheduleJPARepository concertScheduleJPARepository;
+
+    @Override
+    public List<ConcertSchedule> findAvailableConcertSchedules(long concertId, LocalDateTime currentDate) {
+        return concertScheduleJPARepository.findAvailableConcertSchedules(concertId, currentDate);
+    }
+
+}
